refactor(view): util함수 폴더 위치 변경
