+++ src/view/components/FormItem/FormItem.ts
@@ -1,4 +1,4 @@
-import createElementByTag from "../../generateComponent/utils/createElementByTag";
+import createElementByTag from "../../utils/createElementByTag";
 
 class FormItem {
   element = createElementByTag({
